refactor(performance): preserve quality budget
