+++ web/components/dashboard/dashboard-page.ts
@@ -31,12 +31,10 @@ import '../chat/chat-drawer'
 import './filters/filter-dock'
 import './filters/filter-control'
 import { DashboardFilterController } from './filters/filter-controller'
-import type { FilterMutationDetail } from './filters/filter-control'
-import type { FilterOptionsNeededDetail } from './filters/filter-control'
+import type { FilterMutationDetail, FilterOptionsNeededDetail } from './filters/filter-control'
 import './report-canvas'
 import './report-footer'
-import './visual-modal'
-import type { VisualActionDetail } from './visual-modal'
+import { type VisualActionDetail } from './visual-modal'
 import './visualization/host'
 import { DashboardVisualizationSignalDecoder } from './visualization/signal-envelope'
 import {
@@ -739,16 +737,9 @@ class LeapViewDashboardPage extends DatastarLit(LitElement) {
   private get status(): DashboardStatus {
     return this.signal<DashboardStatus>('status', emptyStatus)
   }
-
-  /** Explicitly prepare every visual for screenshot/export tooling. */
   async ensureVisualizationsMounted(): Promise<void> {
-    await this.updateComplete
-    const hosts = Array.from(this.renderRoot.querySelectorAll('lv-visualization-host')) as Array<HTMLElement & {
-      ensureMounted(): Promise<void>
-    }>
-    await Promise.all(hosts.map((host) => host.ensureMounted()))
+    await this.updateComplete; await Promise.all(Array.from(this.renderRoot.querySelectorAll('lv-visualization-host')).map((host) => host.ensureMounted()))
   }
-
   private handleReportZoomState = (event: CustomEvent<{ layout?: unknown }>): void => {
     const layout = event.detail?.layout
     if ((layout === 'desktop' || layout === 'mobile') && layout !== this.reportLayout) {
